Add user-configurable delimieter to cletus_log
Previously, it was hard-coded as a colon.  But this interferes with parsing,
since a colon may often be found in log messages.

This will not address the lack of escaping or proper quoting to ensure that
parsing of the log file will always be accurate - and never confused by
a delimiter within the message.